Memberikan kondisi apakah table assign teacher to classes sudah ada atau belum (migrate sukses)

+++ database/migrations/2024_12_25_080519_create_assign_teacher_to_classes_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('assign_teacher_to_classes', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('class_id')->constrained()->onDelete('cascade');
-            $table->foreignId('subject_id')->constrained()->onDelete('cascade');
-            $table->foreignId('teacher_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('assign_teacher_to_classes')) {
+            Schema::create('assign_teacher_to_classes', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('class_id')->constrained()->onDelete('cascade');
+                $table->foreignId('subject_id')->constrained()->onDelete('cascade');
+                $table->foreignId('teacher_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
